+++ kvs.go
@@ -5,13 +5,10 @@ import (
 	"encoding/gob"
 	"errors"
 	"fmt"
-	"io"
 	"log"
 	"os"
 	"sync"
 	"time"
-
-	"github.com/hashicorp/raft"
 )
 
 var isDebug = os.Getenv("DEBUG") == "true"
@@ -35,6 +32,10 @@ type Pair struct {
 	Expire   Expire
 }
 
+type Transaction struct {
+	Pair []Pair
+}
+
 type Expire struct {
 	Time     time.Time
 	NoExpire bool
@@ -48,9 +49,6 @@ func (e *Expire) Expire(t time.Time) bool {
 }
 
 type KV map[[KeyLimit]byte]*Pair
-
-// check raft.FSM impl
-var _ raft.FSM = &KVS{}
 
 type KVS struct {
 	mtx    sync.RWMutex
@@ -90,6 +88,29 @@ func DecodePair(b []byte) (Pair, error) {
 	return pair, nil
 }
 
+func EncodeTrans(p Pair) ([]byte, error) {
+	b := &bytes.Buffer{}
+	e := gob.NewEncoder(b)
+
+	if err := e.Encode(p); err != nil {
+		return nil, ErrEncode
+	}
+
+	return b.Bytes(), nil
+}
+
+func DecodeTrans(b []byte) (Transaction, error) {
+	var transaction Transaction
+	bv := bytes.NewBuffer(b)
+	d := gob.NewDecoder(bv)
+
+	if err := d.Decode(&transaction); err != nil {
+		return Transaction{}, fmt.Errorf("failed restore: %w", err)
+	}
+
+	return transaction, nil
+}
+
 func cloneKV(kv KV) KV {
 	cloned := KV{}
 
@@ -98,128 +119,4 @@ func cloneKV(kv KV) KV {
 	}
 
 	return cloned
-}
-
-func (f *KVS) Apply(l *raft.Log) interface{} {
-	f.mtx.Lock()
-	defer f.mtx.Unlock()
-	p, err := DecodePair(l.Data)
-	if err != nil {
-		return err
-	}
-
-	// TODO mark it as deleted for performance. Remove from Map when creating snapshot
-	if p.IsDelete {
-		debugLog("delete", *p.Key)
-		delete(f.data, *p.Key)
-		delete(f.expire, *p.Key)
-		return true
-	}
-
-	f.data[*p.Key] = &p
-	if !p.Expire.NoExpire {
-		f.expire[*p.Key] = &p
-	}
-
-	return true
-}
-
-func (f *KVS) Snapshot() (raft.FSMSnapshot, error) {
-	// Make sure that any future calls to f.Apply() don't change the snapshot.
-	return &snapshot{cloneKV(f.data)}, nil
-}
-
-func (f *KVS) Restore(r io.ReadCloser) error {
-	var decodedMap KV
-	d := gob.NewDecoder(r)
-
-	if err := d.Decode(&decodedMap); err != nil {
-		return fmt.Errorf("failed restore: %w", err)
-	}
-
-	return nil
-}
-
-type snapshot struct {
-	data KV
-}
-
-func (s *snapshot) Persist(sink raft.SnapshotSink) error {
-	b := &bytes.Buffer{}
-	e := gob.NewEncoder(b)
-
-	err := e.Encode(s.data)
-	if err != nil {
-		return fmt.Errorf("failed data encode: %w", err)
-	}
-
-	_, err = sink.Write(b.Bytes())
-	if err != nil {
-		_ = sink.Cancel()
-		return fmt.Errorf("sink.Write(): %w", err)
-	}
-	return errors.Unwrap(sink.Close())
-}
-
-func (s *snapshot) Release() {}
-
-type RPCInterface struct {
-	KVS         *KVS
-	Raft        *raft.Raft
-	gcc         chan Pair
-	Environment string
-}
-
-const gcMaxBuffer = 65534
-
-const gcInterval = 500 * time.Millisecond
-
-func NewRPCInterface(kvs *KVS, raft *raft.Raft) *RPCInterface {
-	r := &RPCInterface{
-		KVS:  kvs,
-		Raft: raft,
-		gcc:  make(chan Pair, gcMaxBuffer),
-	}
-	go (func(r *RPCInterface) {
-		debugLog("run gc")
-		ticker := time.NewTicker(gcInterval)
-		for {
-			select {
-			case v := <-r.gcc:
-				v.IsDelete = true
-				e, err := EncodePair(v)
-				if err != nil {
-					log.Println(err)
-				}
-				debugLog("apply delete")
-				_ = r.Raft.Apply(e, time.Second)
-			case <-ticker.C:
-				now := time.Now().UTC()
-				r.KVS.mtx.RLock()
-				for _, pair := range r.KVS.expire {
-					if !pair.Expire.Expire(now) {
-						continue
-					}
-					debugLog("detect expire key", pair)
-					r.gcc <- *pair
-				}
-				r.KVS.mtx.RUnlock()
-			}
-		}
-	})(r)
-	return r
-}
-
-func TTLtoTime(d time.Duration) Expire {
-	switch d.Milliseconds() {
-	default:
-		return Expire{
-			Time: time.Now().UTC().Add(d),
-		}
-	case 0:
-		return Expire{
-			NoExpire: true,
-		}
-	}
-
 }
